fix: change supplierId type from int to long in IqcController and IqcService
- Route constraints and method parameters now use 'long' instead of 'int'
- This matches Supplier.Id type (long) and prevents FindAsync type mismatch errors
- Fixes InvalidCastException when calling GetSupplierScore / UpdateSupplierScore API

+++ backend/Services/IqcService.cs
@@ -895,7 +895,7 @@ public class IqcService
         };
     }
 
-    public async Task<SupplierScoreDto?> UpdateSupplierScore(int supplierId, UpdateSupplierScoreDto dto)
+    public async Task<SupplierScoreDto?> UpdateSupplierScore(long supplierId, UpdateSupplierScoreDto dto)
     {
         var supplier = await _db.Suppliers.FindAsync(supplierId);
         if (supplier == null) return null;
